Transcribe set_font, and the slot table it works through
Slot 0 of every font array is the current font. set_font(n) copies slot
n over it; set_font(0) instead asks which slot the current font came
from, by matching far pointers. Verified against the original.

+++ reconstruct/seg0dff.c
@@ -127,7 +127,7 @@ void game_startup(void)
     }
 
     DGU16(0x52df) = load_font(0x00de);          /* "memofnt8.fnt" */
-    set_font(DGU16(0x52df));
+    set_font((int16_t)DGU16(0x52df));
 
     DGU16(0x52f6) = sub_2367c(0x00eb);          /* "mouse.bmp"   */
     DGU16(0x52f4) = sub_24f72(0x00f5);          /* "cp.bmp"      */
